Start moving providers
