+++ packages/sift/src/js/components/Column.tsx
@@ -1,7 +1,7 @@
 import h from "@macrostrat/hyper";
 import React from "react";
 import { useParams } from "react-router-dom";
-import { Link as VikeLink } from "~/renderer/Link";
+import { Link as VikeLink } from "~/components/navigation/link-component";
 import Chart from "./Chart";
 import Footer from "./Footer";
 import { SiftLink } from "./Link";
@@ -285,13 +285,13 @@ class _Column extends React.Component {
           <PrevalentTaxa data={this.state.prevalentTaxa} />
 
           <StratColumn data={this.state.units} />
-          <VikeLink
+          <LinkComponent
             href={"/columns/" + this.state.properties.col_id}
             target="_blank"
             className="normalize-link alternate-column"
           >
             Alternate column view
-          </VikeLink>
+          </LinkComponent>
         </div>
 
         <Footer data={this.state.refs} />
