+++ src/components/buttons.ts
@@ -1,6 +1,6 @@
 import { Tag, AnchorButton } from "@blueprintjs/core";
 import h from "@macrostrat/hyper";
-import { Link } from "~/renderer/Link";
+import { Link } from "~/components/navigation/link-component";
 
 export function DevLink({ children, tagText = "Beta", ...props }) {
   return h(Link, props, [children, " ", h(Tag, { minimal: true }, [tagText])]);
