+++ src/components/index.ts
@@ -5,4 +5,5 @@ export * from "./tags";
 export * from "./map-info";
 export * from "./lithology";
 export * from "./buttons";
+export * from "./navigation";
 export { buildCrossSectionLayers } from "~/_utils/map-layers";
