+++ src/components/navigation/index.ts
@@ -1,0 +1,2 @@
+export * from "./breadcrumbs";
+export * from "./link-component";
