+++ src/renderer/index.ts
@@ -1,3 +1,3 @@
-export * from "./breadcrumbs";
+export * from "../components/navigation/breadcrumbs";
 export * from "./page-context";
 export * from "./types";
